Refactor core temperature monitoring functionality via ANT+
