+++ apps/coretemp/metadata.json
@@ -1,7 +1,7 @@
 {
   "id": "coretemp",
   "name": "CoreTemp",
-  "version": "0.7",
+  "version": "0.8",
   "description": "Display CoreTemp device sensor data",
   "icon": "coretemp.png",
   "type": "app",
@@ -11,9 +11,15 @@
   "storage": [
     {"name":"coretemp.wid.js","url":"widget.js"},
     {"name":"coretemp.app.js","url":"coretemp.js"},
+    {"name":"coretemp.runtime","url":"runtime.js"},
+    {"name":"coretemp.ble","url":"ble.js"},
+    {"name":"coretemp.protocol","url":"controlpoint.js"},
+    {"name":"coretemp.hrm","url":"hrm.js"},
+    {"name":"coretemp.store","url":"store.js"},
     {"name":"CORESensor","url":"lib.js"},
     {"name":"coretemp.recorder.js","url":"recorder.js"},
     {"name":"coretemp.settings.js","url":"settings.js"},
+    {"name":"coretemp.settingsui","url":"settingsui.js"},
     {"name":"coretemp.img","url":"coretemp-icon.js","evaluate":true},
     {"name":"coretemp.0.boot.js","url":"boot.js"}
   ],
